+++ client/views/index.js
@@ -9,7 +9,6 @@ export { default as contentFilter } from './content-filter.js';
 export { default as context } from './context.js';
 export { default as editors } from './editors.js';
 export { default as expand } from './expand.js';
-export { default as fallback } from './fallback.js';
 export { default as headers } from './headers.js';
 export { default as hstack } from './hstack.js';
 export { default as html } from './html.js';