+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Fixed model assets generation when non-default config file is specified
 - Client
     - Fixed expanded state for `tree` view leafs when `expanded` setting is a number
+    - Improved display of long strings and strings that contains newlines in `struct` view
 
 ## 1.0.0-beta.12 (07-03-2019)